Move the GridMenuPlayer to its own file
